+++ MathsGame/bin/.gitignore
@@ -1,0 +1,2 @@
+/app/
+/module-info.class
