+++ MathsGame/src/app/LoginPage.java
@@ -1,0 +1,7 @@
+package app;
+
+import javax.swing.JFrame;
+
+public class LoginPage {
+
+}
