+++ MathsGame/src/app/Main.java
@@ -2,7 +2,6 @@ package app;
 
 public class Main {
 	public static void main(String[] args) {
-		MainWindow mainWindow= new MainWindow(900,900); 
-	    mainWindow.getFrame().setVisible(true);	    
+		 
 	}
 }
